\name{qsea-package}
\alias{qsea-package}
\alias{qsea}
\alias{QSEA}
\docType{package}
\title{
QSEA: Quantitative sequencing enrichment analysis and visualization
}
\description{
    QSEA (quantitative sequencing enrichment analysis) was developed as the 
    successor of the MEDIPS package for analyzing data derived from 
    methylated DNA immunoprecipitation (MeDIP) experiments followed by 
    sequencing (MeDIP-seq). 
    However, qsea provides functionality for the analysis of 
    other kinds of quantitative sequencing data 
    (e.g. ChIP-seq, MBD-seq, CMS-seq and others) 
    including calculation of differential enrichment between groups of samples.
}
\author{
Matthias Lienhard, Lukas Chavez and Ralf Herwig

Maintainer: Matthias Lienhard <lienhard@molgen.mpg.de>
}
\references{
Lienhard M, Grimm C, Morkel M, Herwig R, Chavez L., (Bioinformatics, 2014): 
MEDIPS: genome-wide differential coverage analysis of sequencing data derived 
from DNA enrichment experiments.
}
\keyword{ package }
